Fix issue when reordering commits.
Before pushing, set base branch to master for all relevant PRs.  This
avoids bustage when reordering PRs.

I think what was happening was,

  - I reorder commits locally.
  - I push the commits, before updating the PRs' base branches.
  - Github notices that certain PRs are now empty and closes them.

For example:

 - Original history: X -> Y
 - PR A: Based on master, ends at branch X.
 - PR B: Based on A, ends at branch Y.

After reordering X and Y:

 - Force-pushed history: Y -> X
 - PR A: Still based on master, still ends at branch X, so now contains
   Y and X!
 - PR B: Still based on X, still ends at branch Y, and notice that there
   are no commits in the range `X..Y`.
 - Therefore github closes PR B.

To fix this, before force-pushing branches, we set all bases to master.
Then we force-push, which is now safe.  Finally we fix the base
branches.