Handle squash-committed PRs better.
Squash-committed PRs put the PR description in the commit message, but
we put git-pr-chain stuff in there, and we don't want this in the commit
message.

Also the `git pull --rebase` could result in spurrious conflicts when we
have a squash merge.  Now we do something which I think will not
conflict.